requirements.txt
google-cloud-bigquery >= 1.24.0
simple-salesforce >= 1.0.0
jsonpickle >= 1.4.1
requests >= 2.23.0
xmltodict >= 0.12.0